requirements.txt
asgiref==3.7.2
coverage==7.3.1
Django==3.2.21
djangorestframework==3.14.0
drf-yasg==1.21.7
inflection==0.5.1
iniconfig==2.0.0
packaging==23.1
pip==23.2.1
pluggy==1.3.0
pytest==7.4.2
pytest-cov==4.1.0
pytest-django==4.5.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
setuptools==65.5.0
sqlparse==0.4.4
uritemplate==4.1.1
